+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add local integration tests for the vSphere provider.
+
 ## [0.13.0] - 2026-07-09
 
 ### Fixed